% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_estimation.R
\name{weak_effect_plot}
\alias{weak_effect_plot}
\title{Weak effect plot}
\usage{
weak_effect_plot(X, M, Y, K, covariates = NULL)
}
\arguments{
\item{X}{Numeric vector. Exposure variable.}

\item{M}{Numeric matrix. Mediators (rows = samples, columns = mediators).}

\item{Y}{Numeric vector. Binary outcome variable (0/1).}

\item{K}{Numeric. Proportion of cases in the population.}

\item{covariates}{Numeric matrix. Covariates included in the model.}
}
\value{
A ggplot2 object showing the histogram of p-values for the mediator-outcome associations.
}
\description{
This function visualizes the distribution of p-values for the association between each mediator and a binary outcome. It highlights potential weak mediator-outcome associations when the mediator is also associated with the exposure.
}
\examples{
data <- data_generate(p = 2000, N = 500, K = 0.05, r = 0.5, res.var = 0.2,
                      pi_11 = 0.09, pi_alpha = 0.3, pi_beta = 0.3,
                      sig2_a = 0.3, sig2_11 = 0.5, sig2_01 = 0.3, seed = 23)
X <- data$x
M <- data$M
Y <- data$y
K <- 0.05

result <- r2_estimation_cf(X, M, Y, K)
print(result)

}
