% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_estimation.R
\name{r2_estimation_cf}
\alias{r2_estimation_cf}
\title{Cross-Fitted R-squared Based Mediation Effect Estimation}
\usage{
r2_estimation_cf(
  X,
  M,
  Y,
  K,
  covariates = NULL,
  nfold = 2,
  fdr_method = "BH",
  fdr_cutoff = 0.01,
  seed = 123
)
}
\arguments{
\item{X}{Numeric vector. Exposure variable.}

\item{M}{Numeric matrix. Mediators (rows = samples, columns = mediators).}

\item{Y}{Numeric vector. Binary outcome variable (0/1).}

\item{K}{Numeric. Proportion of cases in the population.}

\item{covariates}{Numeric matrix. Covariates included in the model.}

\item{nfold}{Integer. Number of cross-fitting folds. Default is 2.}

\item{fdr_method}{Str. FDR control method, can be either 'storey' or any options provided in function p.adjust(). Default is 'BH'.}

\item{fdr_cutoff}{Numeric. FDR cutoff value. Default is 0.01.}

\item{seed}{Integer. Random seed for reproducibility. Default is 123.}
}
\value{
A named list containing:
\describe{
  \item{Rmed.est}{Estimated mediation effect.}
  \item{Q.est}{Estimated Q measure.}
  \item{varl.est}{Total variance estimate.}
  \item{sig2_11.est}{Estimated variance component for indirect effects.}
  \item{sig2_a.est}{Estimated variance component for exposure effects.}
  \item{gamma.est}{Estimated direct effect of exposure on outcome.}
  \item{sig_number}{Number of selected significant mediators.}
  \item{time}{Total computation time in seconds.}
}
}
\description{
This function estimates the mediation effect using cross-fitting.
}
\examples{
data <- data_generate(p = 2000, N = 500, K = 0.05, r = 0.5, res.var = 0.2,
                      pi_11 = 0.09, pi_alpha = 0.3, pi_beta = 0.3,
                      sig2_a = 0.3, sig2_11 = 0.5, sig2_01 = 0.3, seed = 23)
X <- data$x
M <- data$M
Y <- data$y
K <- 0.05

result <- r2_estimation_cf(X, M, Y, K)
print(result)

}
